requirements-hybrid.txt
# ===============================
# 3-Layer Hybrid Contract Analysis
# ===============================

# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Configuration & Validation
pydantic>=1.10.13,<3.0
python-dotenv==1.0.1

# Document Parsing
PyMuPDF==1.24.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.11.0

# NLP / AI (ALL FREE, LOCAL MODELS)
torch>=2.2.1
transformers>=4.37.2
sentence-transformers>=2.3.1
spacy>=3.7.0

# Vector Store (Local, Free)
faiss-cpu>=1.7.4

# Language Detection
langdetect==1.0.9

# Testing
pytest==8.0.0
httpx==0.26.0

# Utilities
numpy>=1.26.0
